Preparing console support
