+++ products/bluebell/core/src/passes/annotate_base_types.rs
@@ -222,7 +222,6 @@ impl IrPass for AnnotateBaseTypes {
         symbol: &mut IrIdentifier,
         symbol_table: &mut SymbolTable,
     ) -> Result<TraversalResult, String> {
-        info!("Testing {:?}", symbol);
         match symbol.kind {
             IrIndentifierKind::Unknown => {
                 if let Some(typeinfo) = self.type_of(symbol, symbol_table) {
@@ -302,7 +301,6 @@ impl IrPass for AnnotateBaseTypes {
                 symbol_table.resolve_qualified_name(&symbol.unresolved, &self.namespace)
             {
                 if symbol_table.is_state(&resolved_name) {
-                    info!("Changing {} to state", resolved_name);
                     symbol.kind = IrIndentifierKind::State;
                 }
             } else {
