+++ products/bluebell/playground/src/app.rs
@@ -1,3 +1,4 @@
+use crate::vm_remote_layout::VmRemoteControlLayout;
 use evm_assembly::function_signature::EvmFunctionSignature;
 use std::collections::HashMap;
 use std::rc::Rc;
@@ -14,7 +15,7 @@ use crate::dropdown::Dropdown;
 use crate::examples::EXAMPLES;
 use crate::machine_view::MachineView;
 use crate::state::{State, StateMessage};
-use crate::vm_remote::VmRemoteControl;
+use crate::vm_remote::VmRemote;
 use gloo_timers::callback::Timeout;
 
 #[derive(Clone, PartialEq)]
@@ -34,17 +35,12 @@ pub struct AppLayoutProps {
 
 pub struct AppLayout {
     props: AppLayoutProps,
-    data: String,
     dispatch: Dispatch<State>,
     state: Rc<State>,
-    function_signature: Option<EvmFunctionSignature>,
-    arguments: Vec<String>,
 }
 
 pub enum AppLayoutMessage {
     UpdateState(Rc<State>),
-    UpdateFunctionSignature(Option<EvmFunctionSignature>),
-    SetArguments(Vec<String>),
 }
 
 impl Component for AppLayout {
@@ -59,11 +55,8 @@ impl Component for AppLayout {
 
         Self {
             props: props.clone(),
-            data: "".to_string(),
             state: dispatch.get(),
             dispatch,
-            function_signature: None,
-            arguments: [].to_vec(),
         }
     }
 
@@ -77,78 +70,12 @@ impl Component for AppLayout {
             AppLayoutMessage::UpdateState(state) => {
                 self.state = state;
             }
-            AppLayoutMessage::UpdateFunctionSignature(signature) => {
-                self.function_signature = signature;
-                let n = if let Some(signature) = &self.function_signature {
-                    signature.arguments.len()
-                } else {
-                    0
-                };
-                self.arguments = vec![String::new(); n];
-            }
-            AppLayoutMessage::SetArguments(args) => {
-                self.arguments = args;
-            }
         }
         true
     }
 
-    fn view(&self, ctx: &Context<Self>) -> Html {
-        let functions = self.state.functions.clone();
-        let functions_clone = self.state.functions.clone();
-        let signature = self.function_signature.clone();
-        let arguments = self.arguments.clone();
-
-        let step_button_click = self.dispatch.apply_callback(|_| StateMessage::RunStep);
-        let eject_button_click = self.dispatch.apply_callback(|_| StateMessage::Reset);
-        let stop_button_click = self.dispatch.apply_callback(|_| StateMessage::Reset);
-        let set_function_signature = ctx
-            .link()
-            .callback(|value| AppLayoutMessage::UpdateFunctionSignature(value));
-        let set_arguments = ctx
-            .link()
-            .callback(|value| AppLayoutMessage::SetArguments(value));
-
+    fn view(&self, _ctx: &Context<Self>) -> Html {
         console::log!(format!("Function load: {}", self.state.function_loaded));
-
-        let run_button_click = {
-            let load_function = !self.state.function_loaded;
-            let maybe_function_signature = signature.clone();
-            let argument_list = arguments.clone();
-            Callback::from(move |_| {
-                if load_function {
-                    let maybe_function_signature = maybe_function_signature.clone();
-                    if let Some(function_signature) = maybe_function_signature {
-                        let function_name = function_signature.name.clone();
-                        let dispatch = Dispatch::<State>::new();
-                        let mut arguments = "".to_string();
-                        for (i, arg) in argument_list.iter().enumerate() {
-                            if i > 0 {
-                                arguments.push_str(",");
-                            }
-                            arguments.push_str(&format!("{}", arg).to_string());
-                        }
-                        dispatch.apply(StateMessage::PrepareFunctionCall {
-                            function_name,
-                            arguments,
-                        });
-                    }
-                } else {
-                    let dispatch = Dispatch::<State>::new();
-                    dispatch.reduce_mut(move |s| {
-                        console::log!("Playing");
-                        s.playing = !s.playing;
-                        if s.playing {
-                            Timeout::new(500, move || {
-                                let dispatch = Dispatch::<State>::new();
-                                dispatch.apply(StateMessage::RunStep);
-                            })
-                            .forget();
-                        }
-                    })
-                }
-            })
-        };
 
         let generate_menu = {
             let selected_view = self.props.view as u32;
@@ -253,122 +180,7 @@ impl Component for AppLayout {
             {
                 if let Some(_) = &self.state.executable {
                     html! {
-                        <VmRemoteControl>
-                            <div class="flex flex-col items-center space-y-4">
-                                <div
-                                    class="p-2 bg-zinc-900 w-full text-gray-100 rounded-md space-x-2 flex flex-col"
-                                >
-                                    <div class="w-full" onmouseover={move |e: MouseEvent| e.stop_propagation()}
-                                     onmousedown={move |e: MouseEvent| e.stop_propagation()}
-                                     onmouseup={move |e: MouseEvent| e.stop_propagation()}>
-                                        <select class="w-full bg-zinc-900 py-2" onchange={move |e: Event| {
-                                            let value = e.target_unchecked_into::<HtmlInputElement>().value();
-                                            if value != "none" {
-                                                let value : usize = value.parse::<usize>().unwrap();
-                                                set_function_signature.emit(functions_clone.get(value).cloned())
-                                            } else {
-                                                set_function_signature.emit(None)
-                                            }
-                                        }}>
-                                            <option value="none" selected={true}>{"(Select function)"}</option>
-                                            {functions.iter().enumerate().map(|(i, v)| html! {
-                                                <option value={i.to_string()}>{v.name.clone()}</option>
-                                            }).collect::<Vec<_>>()}
-                                        </select>
-                                    </div>
-                                    <div class="flex items-center">
-                                        <span class="font-bold">{"PC:"}</span>
-                                        <span>{format!("0x{:02x}", self.state.program_counter)}</span>
-                                        <span>{format!("({})", self.state.program_counter)}</span>
-                                    </div>
-                                </div>
-                                { if let Some(signature) = signature {
-                                    html! {
-                                        <>
-                                        <div class="w-full"  onmouseover={move |e: MouseEvent| e.stop_propagation()}
-                                     onmousedown={move |e: MouseEvent| e.stop_propagation()}
-                                     onmouseup={move |e: MouseEvent| e.stop_propagation()}>
-                                            {signature.arguments.iter().zip(arguments.iter()).enumerate().map(|(i, (t, v))| {
-                                                let arguments = arguments.clone();
-                                                let set_arguments = set_arguments.clone();
-                                                html! {
-                                                    <div>
-                                                        <label>{format!("{:?}", t)}</label>
-                                                        <input key={format!("input{}",i)} value={format!("{}", v)} onkeyup={move |e:KeyboardEvent| {
-                                                                let value = e.target_unchecked_into::<HtmlInputElement>().value();
-                                                                let mut arguments = arguments.clone();
-                                                                arguments[i] = value.clone();
-                                                                set_arguments.emit(arguments);
-                                                                console::log!(format!("Arg {} is '{}'", i, value));
-                                                            }} />
-                                                    </div>
-                                                }
-                                            }).collect::<Vec<_>>()}
-                                        </div>
-                                        <div class="flex items-center space-x-4">
-
-                                            <button
-                                                class="flex items-center justify-center bg-blue-600 text-white w-12 h-12 rounded-full hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 text-center"
-                                                onclick={step_button_click.clone()}
-                                            >
-                                                <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-                                                  <path stroke-linecap="round" stroke-linejoin="round" d="M15 15l6-6m0 0l-6-6m6 6H9a6 6 0 000 12h3" />
-                                                </svg>
-
-                                            </button>
-
-                                            <button
-                                                class="flex items-center justify-center bg-green-600 text-white w-14 h-14 rounded-full hover:bg-green-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-green-500 text-center"
-                                                onclick={run_button_click.clone()}
-                                            >
-                                                {
-                                                    if self.state.playing {
-                                                        html!{
-                                                            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-                                                              <path stroke-linecap="round" stroke-linejoin="round" d="M15.75 5.25v13.5m-7.5-13.5v13.5" />
-                                                            </svg>
-                                                        }
-                                                    } else {
-                                                        html! {
-                                                            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-                                                              <path stroke-linecap="round" stroke-linejoin="round" d="M5.25 5.653c0-.856.917-1.398 1.667-.986l11.54 6.348a1.125 1.125 0 010 1.971l-11.54 6.347a1.125 1.125 0 01-1.667-.985V5.653z" />
-                                                            </svg>
-                                                        }
-                                                    }
-                                                }
-
-                                            </button>
-
-                                            <button
-                                                class="flex items-center justify-center bg-orange-600 text-white w-12 h-12 rounded-full hover:bg-orange-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500 text-center"
-                                                onclick={stop_button_click.clone()}
-                                            >
-                                                <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-                                                  <path stroke-linecap="round" stroke-linejoin="round" d="M5.25 7.5A2.25 2.25 0 017.5 5.25h9a2.25 2.25 0 012.25 2.25v9a2.25 2.25 0 01-2.25 2.25h-9a2.25 2.25 0 01-2.25-2.25v-9z" />
-                                                </svg>
-
-                                            </button>
-
-
-                                            <button
-                                                class="flex items-center justify-center bg-red-600 text-white w-8 h-8 rounded-full hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500 text-center"
-                                                onclick={eject_button_click.clone()}
-                                            >
-                                                <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-                                                  <path stroke-linecap="round" stroke-linejoin="round" d="M5.25 7.5A2.25 2.25 0 017.5 5.25h9a2.25 2.25 0 012.25 2.25v9a2.25 2.25 0 01-2.25 2.25h-9a2.25 2.25 0 01-2.25-2.25v-9z" />
-                                                </svg>
-
-                                            </button>
-
-                                        </div>
-                                        </>
-                                    }
-                                } else {
-                                    html!{}
-                                }}
-
-                            </div>
-                        </VmRemoteControl>
+                        <VmRemote />
 
                     }
                 } else {
